* transcode_data.h (rb_transcoder): resetsize_func and resetstate_func   also returns ssize_t.
* enc/trans/iso2022.trans: follow the type change.



git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@19354 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ transcode_data.h
@@ -87,8 +87,8 @@ struct rb_transcoder {
     ssize_t (*func_io)(void*, VALUE, const unsigned char*, size_t); /* info  -> output */
     ssize_t (*func_so)(void*, const unsigned char*, size_t, unsigned char*, size_t); /* start -> output */
     ssize_t (*finish_func)(void*, unsigned char*, size_t); /* -> output */
-    int (*resetsize_func)(void*); /* -> len */
-    int (*resetstate_func)(void*, unsigned char*, size_t); /* -> output */
+    ssize_t (*resetsize_func)(void*); /* -> len */
+    ssize_t (*resetstate_func)(void*, unsigned char*, size_t); /* -> output */
 };
 
 void rb_declare_transcoder(const char *enc1, const char *enc2, const char *lib);
